Reorganize tests into unit and functional.
- Also enables CI.
