+++ .travis.yml
@@ -1,0 +1,10 @@
+language: php
+
+php:
+  - 7.0
+sudo: false
+
+before_script:
+  - travis_retry composer install --prefer-source --no-interaction
+
+script: composer test
