+++ src/Handler.php
@@ -155,7 +155,7 @@ class Handler {
  * Includes the autoloader created by Composer.
  *
  * This file was generated by composer-scaffold.
- *
+ *.
  * @see composer.json
  * @see index.php
  * @see core/install.php
